improvements

+++ scripts/find_motif.sh
@@ -12,9 +12,9 @@ helpFunction()
 	echo -e "\t-h, --help\n"
 	echo "Example run calls below:"
 	echo ""
-	echo "$0 -i myfasta.fa -w 10 -t FASTA\n"
-	echo "$0 -i mybed.bed -w 15 -t BED\n"
-	echo "$0 -i mygenes.txt -w 8 -t GENES\n"
+	echo "$0 -i myfasta.fa -w 10 -t FASTA"
+	echo "$0 -i mybed.bed -w 10 -t BED"
+	echo "$0 -i mygenes.txt -w 10 -t GENES"
 	exit 1 # Exit script after printing help
 }
 # check for args
@@ -96,10 +96,11 @@ scanSeqs() {
 
 #Remove sequences from FASTA that are not compatabile
 preProcessing() {
-	awk 'NR % 2 == 0 {print}' $1 > "${1::-3}.txt"
-	grep -vE "(X)" "${1::-3}.txt" | grep -vE "(N)" | grep -vE "(n)" | tr '[:upper:]' '[:lower:]' > "${1::-3}_filtered.txt"
-	rm "${1::-3}.txt"
-	mv "${1::-3}_filtered.txt" "${1::-3}.txt"
+	awk 'NR % 2 == 0 {print}' $INPUT > "${INPUT::-3}.txt"
+	grep -vE "(X)" "${INPUT::-3}.txt" | grep -vE "(N)" | grep -vE "(n)" | tr '[:upper:]' '[:lower:]' > "${INPUT::-3}_filtered.txt"
+	rm "${INPUT::-3}.txt"
+	mv "${INPUT::-3}_filtered.txt" "${INPUT::-3}.txt"
+	INPUT="${INPUT::-3}.txt"
 }
 
 ############## WIDTH INFERENCE ######################
@@ -179,6 +180,7 @@ startAnalysis() {
 			python run_gibbs.py -i $INPUT -w $WIDTH -m conquer
 			exit 0
 		else
+			# run parallel if slurm can be used
 			if [ -x "$(command -v sbatch)" ]
 			then
 				parallelRun
@@ -200,7 +202,7 @@ startAnalysis() {
 	then
 		if ! [ -x "$(command -v sbatch)" ] 
 		then
-			echo "Error : parallel job submission not supported, use -s instead of -p"
+			echo "Error : parallel job submission requires Slurm, use -s instead of -p"
 			exit 1
 		else
 			parallelRun
@@ -229,8 +231,6 @@ bedToFasta() {
 		gzip hg38.fa
 		echo "Preprocessing fasta..."
 		preProcessing "${INPUT::-4}.fa"
-		#Redirect input var to processed fasta
-		INPUT="${INPUT::-4}.txt"
 	else
 		echo "Error: bedtools is not installed or is not executable from your path."
 		exit 0
@@ -240,48 +240,50 @@ bedToFasta() {
 ################### GENES TO BED #############################
 
 genesToBed() {
-	LIST=$1
-	#Download GENCODE v33 and hg38 chrom sizes
-	if [ ! -f gencode.v33.annotation.gff3 ]
+	if [ -x "$(command -v bedtools)" ] 
 	then
-		wget -nc ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_33/gencode.v33.annotation.gff3.gz
-		gunzip gencode.v33.annotation.gff3.gz
-	fi
-	if [! -f hg38.chrom.sizes]
-	then
+		#Download GENCODE v33 and hg38 chrom sizes
+		if [ ! -f gencode.v33.annotation.gff3 ]
+		then
+			wget -nc ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_33/gencode.v33.annotation.gff3.gz
+			gunzip gencode.v33.annotation.gff3.gz
+		fi
 		wget -nc http://hgdownload.soe.ucsc.edu/goldenPath/hg38/bigZips/hg38.chrom.sizes
+
+		#GET GENE AND EXON COORDINATES
+		echo "Mapping gene names to gene and exon coordinates..."
+		# returns gene and exon coordinates (2 files) for each gene in the list
+		python ../scripts/get_gene_and_exon_coordinates.py $INPUT
+		GENEBED="${INPUT::-4}_genes.bed"
+		EXONBED="${INPUT::-4}_exons.bed"
+
+		#MERGE, SUBTRACT EXONS FROM GENE, AND FLANK GENE 20KB TO CAPTURE REGULATORY ACTIVITY
+		echo "Merging exons..."
+		sort -k1,1 -k2,2n $EXONBED | bedtools merge -i stdin > "${EXONBED::-4}_merged.bed"
+		EXONMERGEDBED="${EXONBED::-4}_merged.bed"
+		echo "Subtracting merged exons from gene and getting gene flanks..."
+		sort -k 1,1 -k 2,2n $GENEBED > "${GENEBED::-4}_sorted.bed"
+		GENESORTEDBED="${GENEBED::-4}_sorted.bed"
+		bedtools subtract -a $GENESORTEDBED -b $EXONMERGEDBED > "${INPUT::-4}_introns.bed"
+		INTRONBED="${INPUT::-4}_introns.bed"
+		bedtools flank -i $GENESORTEDBED -g hg38.chrom.sizes -b 20000 | bedtools subtract -a stdin -b $GENESORTEDBED > "${GENEBED::-4}_20KBflank.bed"
+		GENEFLANKBED="${GENEBED::-4}_20KBflank.bed"
+
+		#ADD GENE FLANKS TO INTRONS
+		cat $GENEFLANKBED $INTRONBED | sort -k 1,1 -k 2,2n | bedtools merge -i stdin > "${INTRONBED::-4}_and_intergenics.bed"
+		
+		mv "${INPUT::-4}_introns_and_intergenics.bed" "${INPUT::-4}.bed"
+		# set glob var to new format
+		INPUT="${INPUT::-4}.bed"
+
+		#CLEANUP AND COMPLETE
+		gzip ../scripts/gencode.v33.annotation.gff3
+		rm GENEFLANKBED INTRONBED GENESORTEDBED EXONMERGEDBED EXONBED GENEBED
+		echo "Done mapping genes to BED."
+	else
+		echo "Error: bedtools is not installed or is not executable from your path."
+		exit 0
 	fi
-
-	#GET GENE AND EXON COORDINATES
-	echo "Mapping gene names to gene and exon coordinates..."
-	# returns gene and exon coordinates (2 files) for each gene in the list
-	python ../scripts/get_gene_and_exon_coordinates.py $LIST
-	GENEBED="${LIST::-4}_genes.bed"
-	EXONBED="${LIST::-4}_exons.bed"
-
-	#MERGE, SUBTRACT EXONS FROM GENE, AND FLANK GENE 20KB TO CAPTURE REGULATORY ACTIVITY
-	echo "Merging exons..."
-	sort -k1,1 -k2,2n $EXONBED | bedtools merge -i stdin > "${EXONBED::-4}_merged.bed"
-	EXONMERGEDBED="${EXONBED::-4}_merged.bed"
-	echo "Subtracting merged exons from gene and getting gene flanks..."
-	sort -k 1,1 -k 2,2n $GENEBED > "${GENEBED::-4}_sorted.bed"
-	GENESORTEDBED="${GENEBED::-4}_sorted.bed"
-	bedtools subtract -a $GENESORTEDBED -b $EXONMERGEDBED > "${LIST::-4}_introns.bed"
-	INTRONBED="${LIST::-4}_introns.bed"
-	bedtools flank -i $GENESORTEDBED -g hg38.chrom.sizes -b 20000 | bedtools subtract -a stdin -b $GENESORTEDBED > "${GENEBED::-4}_20KBflank.bed"
-	GENEFLANKBED="${GENEBED::-4}_20KBflank.bed"
-	done
-
-	#ADD GENE FLANKS TO INTRONS
-	cat $GENEFLANKBED $INTRONBED | sort -k 1,1 -k 2,2n | bedtools merge -i stdin > "${INTRONBED::-4}_and_intergenics.bed"
-	done
-	mv "${LIST::-4}_introns_and_intergenics.bed" "${INPUT::-4}.bed"
-	# reset glob var INPUT
-	INPUT="${INPUT::-4}.bed"
-	#CLEANUP AND COMPLETE
-	gzip ../scripts/gencode.v33.annotation.gff3
-	rm GENEFLANKBED INTRONBED GENESORTEDBED EXONMERGEDBED EXONBED GENEBED
-	echo "Done mapping genes to BED."
 }
 
 ###################  MAIN ####################################
@@ -293,6 +295,7 @@ then
 		echo "Error: file extension must be .fa"
 		exit 0
 	else
+		preProcessing
 		alignSeqs
 		scanSeqs
 		startAnalysis
@@ -324,11 +327,13 @@ then
 		exit 0
 	fi
 	#Map genes to regulatory coordinates in hg38
-	genesToBed ${INPUT} 
+	genesToBed
 	#Map bed to fasta
 	bedToFasta
-	#Quality check
+	# infer width if not provided
+	alignSeqs
+	# quality check
 	scanSeqs
-	#Run motif discovery
+	# run motif discovery
 	startAnalysis
 fi
